Fix cuda synchronize method
The native function binding for `torch.cuda.synchronize` had a mismatched name,
which resulted in a link error at runtime when the symbol could not be
found. This commit fixes that by renaming the native binding in C# (from
`THSCuda_synchronize`) to `THSTorchCuda_synchronize` as defined in the
associated header file.