Add unit test
Added a primitive unit test for the `torch.cuda.synchronize` function. The test
will fail before this fix and succeed after it. Note that this test does not
check the behavior of the function, as testing CUDA kernel synchronization is
not easy without depending on the characteristics of the underlying compute
device.