Restore before packing
The pack.proj used here was Restoring and Packing in the same MSBuild
evaluation, causing it to not get any of the nuget generated props or
targets.

Fix by making these separate and calling as a separate evaluation.

That requires configuration to be specified since intermediate directory
includes configuration (erroneously).